backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
httpx==0.28.1
sqlmodel==0.0.22
apscheduler==3.11.0
python-dotenv==1.0.1
